pyproject.toml
[project]
name = "erosion-r-and-d"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0",
    "torchvision",
    "segmentation-models-pytorch==0.3.3",
    "timm",
    "huggingface-hub",
    "numpy",
    "pandas",
    "matplotlib",
    "Pillow",
    "streamlit",
    "boto3",
    "pyarrow",
    "duckdb",
    "plotly",
    "pyproj",
    "python-dotenv",
    "tqdm",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
